Fix CMDLINE_OPT_PARAM arguments in plugins failing to load
When the server is called with a plugin-defined argument expecting
additional values, in the form '--foo bar', the parser is unable to
skip the additional value until the plugin is fully loaded.

The issue was caused by the first scans of the argument list to find
CMDLINE_OPT_PREINIT or CMDLINE_OPT_SILENT arguments, which happen before
plugins are loaded, and would cause the parser to choke on the
additional values, interpreting them as separate arguments.

This is related to PR #1552, where we attempted to use this option for
the first time.

Signed-off-by: Haru <haru@dotalux.com>